Refactor validation functions to class
