+++ src/main/java/utility/Parser.java
@@ -3,7 +3,6 @@ package utility;
 import constants.ErrorConstant;
 import constants.HealthConstant;
 import constants.UiConstant;
-import constants.WorkoutConstant;
 import ui.Output;
 
 import health.Appointment;
@@ -58,29 +57,6 @@ public class Parser {
     }
 
     /**
-     * Validates that the input date string is correctly formatted in DD-MM-YYYY.
-     *
-     * @param date The string date from user input.
-     * @throws CustomExceptions.InvalidInput If there are invalid date inputs.
-     */
-    public static void validateDateInput(String date) throws CustomExceptions.InvalidInput {
-        if (!date.matches(UiConstant.VALID_DATE_REGEX)) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.INVALID_DATE_ERROR);
-        }
-        String[] parts = date.split(UiConstant.DASH);
-        int day = Integer.parseInt(parts[0]);
-        int month = Integer.parseInt(parts[1]);
-
-        if (day < UiConstant.MIN_DAY || day > UiConstant.MAX_DAY) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.INVALID_DAY_ERROR);
-        }
-
-        if (month < UiConstant.MIN_MONTH || month > UiConstant.MAX_MONTH) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.INVALID_MONTH_ERROR);
-        }
-    }
-
-    /**
      * Splits user input for Delete command into item and index.
      *
      * @param input A user-provided string.
@@ -98,25 +74,6 @@ public class Parser {
     }
 
     /**
-     * Validates the delete input details.
-     *
-     * @param deleteDetails A list containing the details for the delete command.
-     * @throws CustomExceptions.InvalidInput If the details specified are invalid.
-     * @throws CustomExceptions.InsufficientInput If empty strings are used.
-     */
-    public static void validateDeleteInput(String[] deleteDetails) throws CustomExceptions.InvalidInput,
-            CustomExceptions.InsufficientInput {
-        if (deleteDetails[0].isEmpty() || deleteDetails[1].isEmpty()) {
-            throw new CustomExceptions.InsufficientInput(ErrorConstant.INSUFFICIENT_DELETE_PARAMETERS_ERROR);
-        }
-        validateFilter(deleteDetails[0].toLowerCase());
-
-        if (!deleteDetails[1].matches(UiConstant.VALID_POSITIVE_INTEGER_REGEX)) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.INVALID_INDEX_ERROR);
-        }
-    }
-
-    /**
      * Parses and validates user input for the delete command. Returns a list of parsed user input
      * containing the filter string and the index to delete.
      *
@@ -126,34 +83,13 @@ public class Parser {
     public static String[] parseDeleteInput(String userInput) {
         try {
             String[] deleteDetails = splitDeleteInput(userInput);
-            validateDeleteInput(deleteDetails);
+            Validation.validateDeleteInput(deleteDetails);
             return deleteDetails;
         } catch (CustomExceptions.InvalidInput | CustomExceptions.InsufficientInput e) {
             Output.printException(e.getMessage());
             return null;
         }
 
-    }
-
-    /**
-     * Validates whether the filter string is either 'run', 'gym', 'bmi', 'period' 
-     * or 'appointment'. 
-     * 
-     * @param filter The filter string to be checked.
-     * @throws CustomExceptions.InvalidInput If the filter string is none of them.
-     */
-    public static void validateFilter (String filter) throws CustomExceptions.InvalidInput {
-        if (filter.equals(WorkoutConstant.RUN)
-                || filter.equals(WorkoutConstant.GYM) 
-                || filter.equals(HealthConstant.BMI)
-                || filter.equals(HealthConstant.PERIOD)
-                || filter.equals(HealthConstant.APPOINTMENT)
-                || filter.equals(WorkoutConstant.ALL)) {
-            return;
-        }
-        throw new CustomExceptions.InvalidInput(ErrorConstant.INVALID_ITEM_ERROR
-                + System.lineSeparator()
-                + ErrorConstant.CORRECT_FILTER_ITEM_FORMAT);
     }
 
     //@@author JustinSoh
@@ -170,7 +106,7 @@ public class Parser {
             if (type.isBlank()) {
                 throw new CustomExceptions.InsufficientInput(ErrorConstant.INVALID_HISTORY_FILTER_ERROR);
             }
-            validateFilter(type.toLowerCase());
+            Validation.validateFilter(type.toLowerCase());
             return type.toLowerCase();
         } catch (CustomExceptions.InvalidInput | CustomExceptions.InsufficientInput e) {
             Output.printException(e.getMessage());
@@ -186,36 +122,10 @@ public class Parser {
     public static void parseBmiInput(String userInput) throws CustomExceptions.InvalidInput,
             CustomExceptions.InsufficientInput {
         String[] bmiDetails = splitBmiInput(userInput);
-        validateBmiInput(bmiDetails);
+        Validation.validateBmiInput(bmiDetails);
         Bmi newBmi = new Bmi(bmiDetails[0], bmiDetails[1], bmiDetails[2]);
         HealthList.addBmi(newBmi);
         Output.printAddBmi(newBmi);
-    }
-
-    /**
-     * Validates Bmi details entered.
-     *
-     * @param bmiDetails List of strings representing BMI details.
-     * @throws CustomExceptions.InvalidInput If there are any errors in the details entered.
-     */
-    public static void validateBmiInput(String[] bmiDetails) throws CustomExceptions.InvalidInput,
-            CustomExceptions.InsufficientInput {
-        if (bmiDetails[0].isEmpty()
-                || bmiDetails[1].isEmpty()
-                || bmiDetails[2].isEmpty()) {
-            throw new CustomExceptions.InsufficientInput(ErrorConstant.INSUFFICIENT_BMI_PARAMETERS_ERROR);
-        }
-
-        if (!bmiDetails[0].matches(UiConstant.VALID_TWO_DP_NUMBER_REGEX) ||
-                !bmiDetails[1].matches(UiConstant.VALID_TWO_DP_NUMBER_REGEX)) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.HEIGHT_WEIGHT_INPUT_ERROR);
-        }
-        validateDateInput(bmiDetails[2]);
-        LocalDate date = parseDate(bmiDetails[2]);
-        if (date.isAfter(LocalDate.now())) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.DATE_IN_FUTURE_ERROR);
-        }
-
     }
 
     //@@author syj02
@@ -249,7 +159,7 @@ public class Parser {
     public static void parsePeriodInput(String userInput) throws CustomExceptions.InvalidInput,
             CustomExceptions.InsufficientInput {
         String[] periodDetails = splitPeriodInput(userInput);
-        validatePeriodInput(periodDetails);
+        Validation.validatePeriodInput(periodDetails);
         Period newPeriod = new Period(periodDetails[0], periodDetails[1]);
         HealthList.addPeriod(newPeriod);
         Output.printAddPeriod(newPeriod);
@@ -275,43 +185,6 @@ public class Parser {
     }
 
     /**
-     * Validates Period details entered.
-     *
-     * @param periodDetails List of strings representing Period details.
-     * @throws CustomExceptions.InvalidInput If there are any errors in the details entered.
-     */
-    public static void validatePeriodInput(String[] periodDetails) throws CustomExceptions.InvalidInput,
-            CustomExceptions.InsufficientInput {
-        if (periodDetails[0].isEmpty() || periodDetails[1].isEmpty()) {
-            throw new CustomExceptions.InsufficientInput(ErrorConstant.INSUFFICIENT_PERIOD_PARAMETERS_ERROR);
-        }
-
-        try {
-            validateDateInput(periodDetails[0]);
-        } catch (CustomExceptions.InvalidInput e) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.INVALID_START_DATE_ERROR
-                    + System.lineSeparator()
-                    + e.getMessage());
-        }
-        try {
-            validateDateInput(periodDetails[1]);
-        } catch (CustomExceptions.InvalidInput e) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.INVALID_END_DATE_ERROR
-                    + System.lineSeparator()
-                    + e.getMessage());
-        }
-
-        LocalDate startDate = parseDate(periodDetails[0]);
-        LocalDate endDate = parseDate(periodDetails[1]);
-        if (startDate.isAfter(LocalDate.now())) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.START_DATE_IN_FUTURE_ERROR);
-        }
-        if (startDate.isAfter(endDate)) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.PERIOD_END_BEFORE_START_ERROR);
-        }
-    }
-
-    /**
      * Parses input for Prediction command.
      * Prints period prediction if possible.
      *
@@ -324,90 +197,6 @@ public class Parser {
             Period.printNextCyclePrediction(nextPeriodStartDate);
         } else {
             throw new CustomExceptions.InsufficientInput(ErrorConstant.UNABLE_TO_MAKE_PREDICTIONS_ERROR);
-        }
-    }
-
-    /**
-     * Validates the time used in HH:MM format.
-     *
-     * @param time String representing the time to check.
-     * @throws CustomExceptions.InvalidInput If time is formatted wrongly.
-     */
-    public static void validateTimeInput(String time) throws CustomExceptions.InvalidInput {
-        if (!time.matches(UiConstant.VALID_TIME_REGEX)) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.INVALID_TIME_ERROR);
-        }
-        String [] parts = time.split(UiConstant.SPLIT_BY_COLON);
-        int hours = Integer.parseInt(parts[0]);
-        int minutes = Integer.parseInt(parts[1]);
-
-        if (hours < UiConstant.MIN_HOURS || hours > UiConstant.MAX_HOURS) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.INVALID_HOURS_ERROR);
-        }
-        if (minutes < UiConstant.MIN_MINUTES || minutes > UiConstant.MAX_MINUTES) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.INVALID_MINUTES_ERROR);
-        }
-    }
-
-    /**
-     * Validates the time used in HH:MM format.
-     *
-     * @param time String representing the time to check.
-     * @throws CustomExceptions.InvalidInput If time is formatted wrongly.
-     */
-    public static void validateRunTimeInput(String time) throws CustomExceptions.InvalidInput {
-        if (!time.matches(UiConstant.VALID_TIME_REGEX) &&
-                !time.matches(UiConstant.VALID_TIME_WITH_HOURS_REGEX)) {
-            throw new CustomExceptions.InvalidInput("Invalid time format. " +
-                    "Format is HH:MM:SS or MM:SS with integers");
-        }
-        String [] parts = time.split(":");
-        int hours = -1; // if not needed, leave as -1.
-        int minutes;
-        int seconds;
-
-        if (parts.length == 2) {
-            minutes = Integer.parseInt(parts[0]);
-            seconds = Integer.parseInt(parts[1]);
-        } else if (parts.length == 3) {
-            hours = Integer.parseInt(parts[0]);
-            minutes = Integer.parseInt(parts[1]);
-            seconds = Integer.parseInt(parts[2]);
-        } else {
-            throw new CustomExceptions.InvalidInput("Invalid time format. Format is HH:MM:SS or MM:SS with integers");
-        }
-        if (minutes < 1 || minutes > 60) {
-            throw new CustomExceptions.InvalidInput("Minutes must be a positive integer between 01 and 59.");
-        }
-
-        if (seconds < 1 || seconds > 60) {
-            throw new CustomExceptions.InvalidInput("Minutes must be a positive integer between 01 and 59.");
-        }
-
-        if (hours == 0) {
-            throw new CustomExceptions.InvalidInput("Hours cannot be 0. Use MM:SS instead");
-        }
-    }
-
-    /**
-     * Validates Appointment details entered.
-     *
-     * @param appointmentDetails List of strings representing Appointment details.
-     * @throws CustomExceptions.InvalidInput If there are any errors in the details entered.
-     */
-    public static void validateAppointmentDetails(String[] appointmentDetails)
-            throws CustomExceptions.InvalidInput, CustomExceptions.InsufficientInput {
-        if (appointmentDetails[0].isEmpty()
-                || appointmentDetails[1].isEmpty()
-                || appointmentDetails[2].isEmpty()) {
-            throw new CustomExceptions.InsufficientInput( ErrorConstant
-                    .INSUFFICIENT_APPOINTMENT_PARAMETERS_ERROR);
-        }
-        validateDateInput(appointmentDetails[0]);
-        validateTimeInput(appointmentDetails[1]);
-
-        if (appointmentDetails[2].length() > HealthConstant.MAX_DESCRIPTION_LENGTH) {
-            throw new CustomExceptions.InvalidInput(ErrorConstant.DESCRIPTION_LENGTH_ERROR);
         }
     }
 
@@ -440,7 +229,7 @@ public class Parser {
     public static void parseAppointmentInput(String userInput) throws CustomExceptions.InvalidInput,
             CustomExceptions.InsufficientInput {
         String[] appointmentDetails = splitAppointmentDetails(userInput);
-        validateAppointmentDetails(appointmentDetails);
+        Validation.validateAppointmentDetails(appointmentDetails);
         Appointment newAppointment = new Appointment(appointmentDetails[0],
                 appointmentDetails[1],
                 appointmentDetails[2]);
@@ -460,7 +249,6 @@ public class Parser {
         if (index == -1 || index == input.length() - delimiter.length()) {
             return "";
         }
-
         int startIndex = index + delimiter.length();
         int endIndex = input.indexOf("/", startIndex);
         if (endIndex == -1) {
